Canvas - Changed frame_index to 16-bit.

+++ src/canvas/canvas.cpp
@@ -51,7 +51,7 @@ namespace PixelMaestro {
 	 * @param radius The circle's radius.
 	 * @param fill Whether to fill the circle or leave it empty.
 	 */
-	void Canvas::draw_circle(uint8_t frame_index, uint8_t color_index, uint16_t origin_x, uint16_t origin_y, uint16_t radius, bool fill) {
+	void Canvas::draw_circle(uint16_t frame_index, uint8_t color_index, uint16_t origin_x, uint16_t origin_y, uint16_t radius, bool fill) {
 		// (x – h)^2 + (y – k)^2 = r^2
 		// r = radius, h = origin_x, k = origin_y
 
@@ -89,7 +89,7 @@ namespace PixelMaestro {
 	 * @param size_x Frame width.
 	 * @param size_y Frame height.
 	 */
-	void Canvas::draw_frame(uint8_t frame_index, uint8_t* frame, uint16_t size_x, uint16_t size_y) {
+	void Canvas::draw_frame(uint16_t frame_index, uint8_t* frame, uint16_t size_x, uint16_t size_y) {
 		clear();
 		Point frame_bounds = Point(size_x, size_y);
 		for (uint16_t y = 0; y < size_y; y++) {
@@ -108,7 +108,7 @@ namespace PixelMaestro {
 	 * @param target_x Ending point x coordinate.
 	 * @param target_y Ending point y coordinate.
 	 */
-	void Canvas::draw_line(uint8_t frame_index, uint8_t color_index, uint16_t origin_x, uint16_t origin_y, uint16_t target_x, uint16_t target_y) {
+	void Canvas::draw_line(uint16_t frame_index, uint8_t color_index, uint16_t origin_x, uint16_t origin_y, uint16_t target_x, uint16_t target_y) {
 		// Calculate slope
 		float slope;
 		if (target_x == origin_x) {
@@ -165,7 +165,7 @@ namespace PixelMaestro {
 	 * @param cursor_x Starting point x coordinate.
 	 * @param cursor_y Starting point y coordinate.
 	 */
-	void Canvas::draw_point(uint8_t frame_index, uint8_t color_index, uint16_t x, uint16_t y) {
+	void Canvas::draw_point(uint16_t frame_index, uint8_t color_index, uint16_t x, uint16_t y) {
 		if (section_.get_dimensions().in_bounds(x, y)) {
 			frames_[frame_index][section_.get_dimensions().get_inline_index(x, y)] = color_index;
 		}
@@ -179,7 +179,7 @@ namespace PixelMaestro {
 	 * @param size_y Height of the rectangle.
 	 * @param fill Whether to fill the rectangle or leave it empty.
 	 */
-	void Canvas::draw_rect(uint8_t frame_index, uint8_t color_index, uint16_t origin_x, uint16_t origin_y, uint16_t size_x, uint16_t size_y, bool fill) {
+	void Canvas::draw_rect(uint16_t frame_index, uint8_t color_index, uint16_t origin_x, uint16_t origin_y, uint16_t size_x, uint16_t size_y, bool fill) {
 		Point cursor = { origin_x, origin_y };
 		for (uint16_t column = 0; column < size_x; column++) {
 			// (Re-)Initialize cursor coordinates.
@@ -214,7 +214,7 @@ namespace PixelMaestro {
 	 * @param font The Font to draw the text in.
 	 * @param text The string to draw.
 	 */
-	void Canvas::draw_text(uint8_t frame_index, uint8_t color_index, uint16_t origin_x, uint16_t origin_y, Font& font, const char* text, uint8_t num_chars) {
+	void Canvas::draw_text(uint16_t frame_index, uint8_t color_index, uint16_t origin_x, uint16_t origin_y, Font& font, const char* text, uint8_t num_chars) {
 		Point cursor = {origin_x, origin_y};
 
 		for (uint16_t letter = 0; letter < num_chars; letter++) {
@@ -250,7 +250,7 @@ namespace PixelMaestro {
 	 * @param point_c_y Third point y-coordinate.
 	 * @param fill Whether to fill the triangle or leave it empty.
 	 */
-	void Canvas::draw_triangle(uint8_t frame_index, uint8_t color_index, uint16_t point_a_x, uint16_t point_a_y, uint16_t point_b_x, uint16_t point_b_y, uint16_t point_c_x, uint16_t point_c_y, bool fill) {
+	void Canvas::draw_triangle(uint16_t frame_index, uint8_t color_index, uint16_t point_a_x, uint16_t point_a_y, uint16_t point_b_x, uint16_t point_b_y, uint16_t point_c_x, uint16_t point_c_y, bool fill) {
 		this->draw_line(frame_index, color_index, point_a_x, point_a_y, point_b_x, point_b_y);
 		this->draw_line(frame_index, color_index, point_b_x, point_b_y, point_c_x, point_c_y);
 		this->draw_line(frame_index, color_index, point_c_x, point_c_y, point_a_x, point_a_y);
@@ -299,7 +299,7 @@ namespace PixelMaestro {
 	 * @param x X coordinate.
 	 * @param y Y coordinate.
 	 */
-	void Canvas::erase_point(uint8_t frame_index, uint16_t x, uint16_t y) {
+	void Canvas::erase_point(uint16_t frame_index, uint16_t x, uint16_t y) {
 		frames_[frame_index][section_.get_dimensions().get_inline_index(x, y)] = 255;
 	}
